[FEAT] FeedLocation 수정 : 위도와 경도 데이터 타입 double로 변경

+++ src/main/java/com/example/trip/domain/FeedLocation.java
@@ -22,9 +22,9 @@ public class FeedLocation {
     @Column(name = "location_name")
     private String name;
 
-    private Long latitude;
+    private double latitude;
 
-    private Long longitude;
+    private double longitude;
 
     private String placeAddress;
